+++ src/spatialdata/_core/query/relational_query.py
@@ -1,7 +1,12 @@
 from __future__ import annotations
 
+import math
+import warnings
+from collections import defaultdict
 from dataclasses import dataclass
-from typing import Any
+from enum import Enum
+from functools import partial
+from typing import Any, Literal
 
 import dask.array as da
 import numpy as np
@@ -21,6 +26,7 @@ from spatialdata.models import (
     SpatialElement,
     TableModel,
     get_model,
+    get_table_keys,
 )
 
 
@@ -49,6 +55,21 @@ def _filter_table_by_element_names(table: AnnData | None, element_names: str | l
     return table
 
 
+def _get_unique_label_values_as_index(element: SpatialElement) -> pd.Index:
+    if isinstance(element, SpatialImage):
+        # get unique labels value (including 0 if present)
+        instances = da.unique(element.data).compute()
+    else:
+        assert isinstance(element, MultiscaleSpatialImage)
+        v = element["scale0"].values()
+        assert len(v) == 1
+        xdata = next(iter(v))
+        # can be slow
+        instances = da.unique(xdata.data).compute()
+    return pd.Index(np.sort(instances))
+
+
+# TODO: replace function use throughout repo by `join_sdata_spatialelement_table`
 def _filter_table_by_elements(
     table: AnnData | None, elements_dict: dict[str, dict[str, Any]], match_rows: bool = False
 ) -> AnnData | None:
@@ -129,6 +150,389 @@ def _filter_table_by_elements(
     return table
 
 
+def _get_joined_table_indices(
+    joined_indices: pd.Index | None,
+    element_indices: pd.RangeIndex,
+    table_instance_key_column: pd.Series,
+    match_rows: Literal["left", "no", "right"],
+) -> pd.Index:
+    """
+    Get indices of the table that are present in element_indices.
+
+    Parameters
+    ----------
+    joined_indices
+        Current indices that have been found to match indices of an element
+    element_indices
+        Element indices to match against table_instance_key_column.
+    table_instance_key_column
+        The column of a table containing the instance ids.
+    match_rows
+        Whether to match the indices of the element and table and if so how. If left, element_indices take priority and
+        if right table instance ids take priority.
+
+    Returns
+    -------
+        The indices that of the table that match the SpatialElement indices.
+    """
+    mask = table_instance_key_column.isin(element_indices)
+    if joined_indices is None:
+        if match_rows == "left":
+            joined_indices = _match_rows(table_instance_key_column, mask, element_indices, match_rows)
+        else:
+            joined_indices = table_instance_key_column[mask].index
+    else:
+        if match_rows == "left":
+            add_indices = _match_rows(table_instance_key_column, mask, element_indices, match_rows)
+            joined_indices = joined_indices.append(add_indices)
+        # in place append does not work with pd.Index
+        else:
+            joined_indices = joined_indices.append(table_instance_key_column[mask].index)
+    return joined_indices
+
+
+def _get_masked_element(
+    element_indices: pd.RangeIndex,
+    element: SpatialElement,
+    table_instance_key_column: pd.Series,
+    match_rows: Literal["left", "no", "right"],
+) -> SpatialElement:
+    """
+    Get element rows matching the instance ids in the table_instance_key_column.
+
+    Parameters
+    ----------
+    element_indices
+        The indices of an element.
+    element
+        The spatial element to be masked.
+    table_instance_key_column
+        The column of a table containing the instance ids
+    match_rows
+         Whether to match the indices of the element and table and if so how. If left, element_indices take priority and
+        if right table instance ids take priority.
+
+    Returns
+    -------
+    The masked spatial element based on the provided indices and match rows.
+    """
+    mask = table_instance_key_column.isin(element_indices)
+    masked_table_instance_key_column = table_instance_key_column[mask]
+    mask_values = mask_values if len(mask_values := masked_table_instance_key_column.values) != 0 else None
+    if match_rows == "right":
+        mask_values = _match_rows(table_instance_key_column, mask, element_indices, match_rows)
+
+    return element.loc[mask_values, :]
+
+
+def _right_exclusive_join_spatialelement_table(
+    element_dict: dict[str, dict[str, Any]], table: AnnData, match_rows: Literal["left", "no", "right"]
+) -> tuple[dict[str, Any], AnnData | None]:
+    regions, region_column_name, instance_key = get_table_keys(table)
+    groups_df = table.obs.groupby(by=region_column_name)
+    mask = []
+    for element_type, name_element in element_dict.items():
+        for name, element in name_element.items():
+            if name in regions:
+                group_df = groups_df.get_group(name)
+                table_instance_key_column = group_df[instance_key]
+                if element_type in ["points", "shapes"]:
+                    element_indices = element.index
+                else:
+                    element_indices = _get_unique_label_values_as_index(element)
+
+                element_dict[element_type][name] = None
+                submask = ~table_instance_key_column.isin(element_indices)
+                mask.append(submask)
+            else:
+                warnings.warn(
+                    f"The element `{name}` is not annotated by the table. Skipping", UserWarning, stacklevel=2
+                )
+                element_dict[element_type][name] = None
+                continue
+
+    if len(mask) != 0:
+        mask = pd.concat(mask)
+        exclusive_table = table[mask, :].copy() if mask.sum() != 0 else None  # type: ignore[attr-defined]
+    else:
+        exclusive_table = None
+
+    return element_dict, exclusive_table
+
+
+def _right_join_spatialelement_table(
+    element_dict: dict[str, dict[str, Any]], table: AnnData, match_rows: Literal["left", "no", "right"]
+) -> tuple[dict[str, Any], AnnData]:
+    if match_rows == "left":
+        warnings.warn("Matching rows ``'left'`` is not supported for ``'right'`` join.", UserWarning, stacklevel=2)
+    regions, region_column_name, instance_key = get_table_keys(table)
+    groups_df = table.obs.groupby(by=region_column_name)
+    for element_type, name_element in element_dict.items():
+        for name, element in name_element.items():
+            if name in regions:
+                group_df = groups_df.get_group(name)
+                table_instance_key_column = group_df[instance_key]
+                if element_type in ["points", "shapes"]:
+                    element_indices = element.index
+                else:
+                    warnings.warn(
+                        f"Element type `labels` not supported for left exclusive join. Skipping `{name}`",
+                        UserWarning,
+                        stacklevel=2,
+                    )
+                    continue
+
+                masked_element = _get_masked_element(element_indices, element, table_instance_key_column, match_rows)
+                element_dict[element_type][name] = masked_element
+            else:
+                warnings.warn(
+                    f"The element `{name}` is not annotated by the table. Skipping", UserWarning, stacklevel=2
+                )
+                continue
+    return element_dict, table
+
+
+def _inner_join_spatialelement_table(
+    element_dict: dict[str, dict[str, Any]], table: AnnData, match_rows: Literal["left", "no", "right"]
+) -> tuple[dict[str, Any], AnnData]:
+    regions, region_column_name, instance_key = get_table_keys(table)
+    groups_df = table.obs.groupby(by=region_column_name)
+    joined_indices = None
+    for element_type, name_element in element_dict.items():
+        for name, element in name_element.items():
+            if name in regions:
+                group_df = groups_df.get_group(name)
+                table_instance_key_column = group_df[instance_key]  # This is always a series
+                if element_type in ["points", "shapes"]:
+                    element_indices = element.index
+                else:
+                    warnings.warn(
+                        f"Element type `labels` not supported for left exclusive join. Skipping `{name}`",
+                        UserWarning,
+                        stacklevel=2,
+                    )
+                    continue
+
+                masked_element = _get_masked_element(element_indices, element, table_instance_key_column, match_rows)
+                element_dict[element_type][name] = masked_element
+
+                joined_indices = _get_joined_table_indices(
+                    joined_indices, element_indices, table_instance_key_column, match_rows
+                )
+            else:
+                warnings.warn(
+                    f"The element `{name}` is not annotated by the table. Skipping", UserWarning, stacklevel=2
+                )
+                element_dict[element_type][name] = None
+                continue
+
+    joined_table = table[joined_indices, :].copy() if joined_indices is not None else None
+    return element_dict, joined_table
+
+
+def _left_exclusive_join_spatialelement_table(
+    element_dict: dict[str, dict[str, Any]], table: AnnData, match_rows: Literal["left", "no", "right"]
+) -> tuple[dict[str, Any], AnnData | None]:
+    regions, region_column_name, instance_key = get_table_keys(table)
+    groups_df = table.obs.groupby(by=region_column_name)
+    for element_type, name_element in element_dict.items():
+        for name, element in name_element.items():
+            if name in regions:
+                group_df = groups_df.get_group(name)
+                table_instance_key_column = group_df[instance_key]
+                if element_type in ["points", "shapes"]:
+                    mask = np.full(len(element), True, dtype=bool)
+                    mask[table_instance_key_column.values] = False
+                    masked_element = element.loc[mask, :] if mask.sum() != 0 else None
+                    element_dict[element_type][name] = masked_element
+                else:
+                    warnings.warn(
+                        f"Element type `labels` not supported for left exclusive join. Skipping `{name}`",
+                        UserWarning,
+                        stacklevel=2,
+                    )
+                    continue
+            else:
+                warnings.warn(
+                    f"The element `{name}` is not annotated by the table. Skipping", UserWarning, stacklevel=2
+                )
+                continue
+
+    return element_dict, None
+
+
+def _left_join_spatialelement_table(
+    element_dict: dict[str, dict[str, Any]], table: AnnData, match_rows: Literal["left", "no", "right"]
+) -> tuple[dict[str, Any], AnnData]:
+    if match_rows == "right":
+        warnings.warn("Matching rows ``'right'`` is not supported for ``'left'`` join.", UserWarning, stacklevel=2)
+    regions, region_column_name, instance_key = get_table_keys(table)
+    groups_df = table.obs.groupby(by=region_column_name)
+    joined_indices = None
+    for element_type, name_element in element_dict.items():
+        for name, element in name_element.items():
+            if name in regions:
+                group_df = groups_df.get_group(name)
+                table_instance_key_column = group_df[instance_key]  # This is always a series
+                if element_type in ["points", "shapes"]:
+                    element_indices = element.index
+                else:
+                    element_indices = _get_unique_label_values_as_index(element)
+
+                joined_indices = _get_joined_table_indices(
+                    joined_indices, element_indices, table_instance_key_column, match_rows
+                )
+            else:
+                warnings.warn(
+                    f"The element `{name}` is not annotated by the table. Skipping", UserWarning, stacklevel=2
+                )
+                continue
+
+    joined_indices = joined_indices.dropna() if joined_indices is not None else None
+    joined_table = table[joined_indices, :].copy() if joined_indices is not None else None
+
+    return element_dict, joined_table
+
+
+def _match_rows(
+    table_instance_key_column: pd.Series,
+    mask: pd.Series,
+    element_indices: pd.RangeIndex,
+    match_rows: str,
+) -> pd.Index:
+    instance_id_df = pd.DataFrame(
+        {"instance_id": table_instance_key_column[mask].values, "index_right": table_instance_key_column[mask].index}
+    )
+    element_index_df = pd.DataFrame({"index_left": element_indices})
+    index_col = "index_left" if match_rows == "right" else "index_right"
+
+    merged_df = pd.merge(
+        element_index_df, instance_id_df, left_on="index_left", right_on="instance_id", how=match_rows
+    )[index_col]
+
+    # With labels it can be that index 0 is NaN
+    if isinstance(merged_df.iloc[0], float) and math.isnan(merged_df.iloc[0]):
+        merged_df = merged_df.iloc[1:]
+
+    return pd.Index(merged_df)
+
+
+class JoinTypes(Enum):
+    """Available join types for matching elements to tables and vice versa."""
+
+    left = partial(_left_join_spatialelement_table)
+    left_exclusive = partial(_left_exclusive_join_spatialelement_table)
+    inner = partial(_inner_join_spatialelement_table)
+    right = partial(_right_join_spatialelement_table)
+    right_exclusive = partial(_right_exclusive_join_spatialelement_table)
+
+    def __call__(self, *args: Any) -> tuple[dict[str, Any], AnnData]:
+        return self.value(*args)
+
+
+class MatchTypes(Enum):
+    """Available match types for matching rows of elements and tables."""
+
+    left = "left"
+    right = "right"
+    no = "no"
+
+
+def join_sdata_spatialelement_table(
+    sdata: SpatialData,
+    spatial_element_name: str | list[str],
+    table_name: str,
+    how: str = "left",
+    match_rows: Literal["no", "left", "right"] = "no",
+) -> tuple[dict[str, Any], AnnData]:
+    """Join SpatialElement(s) and table together in SQL like manner.
+
+    The function allows the user to perform SQL like joins of SpatialElements and a table. The elements are not
+    returned together in one dataframe like structure, but instead filtered elements are returned. To determine matches,
+    for the SpatialElement the index is used and for the table the region key column and instance key column. The
+    elements are not overwritten in the `SpatialData` object.
+
+    The following joins are supported: ``'left'``, ``'left_exclusive'``, ``'inner'``, ``'right'`` and
+    ``'right_exclusive'``. In case of a ``'left'`` join the SpatialElements are returned in a dictionary as is
+    while the table is filtered to only include matching rows. In case of ``'left_exclusive'`` join None is returned
+    for table while the SpatialElements returned are filtered to only include indices not present in the table. The
+    cases for ``'right'`` joins are symmetric to the ``'left'`` joins. In case of an ``'inner'`` join of
+    SpatialElement(s) and a table, for each an element is returned only containing the rows that are present in
+    both the SpatialElement and table.
+
+    For Points and Shapes elements every valid join for argument how is supported. For Labels elements only
+     the ``'left'`` and ``'right_exclusive'`` joins are supported.
+
+    Parameters
+    ----------
+    sdata
+        The SpatialData object containing the tables and spatial elements.
+    spatial_element_name
+        The name(s) of the spatial elements to be joined with the table.
+    table_name
+        The name of the table to join with the spatial elements.
+    how
+        The type of SQL like join to perform, default is ``'left'``. Options are ``'left'``, ``'left_exclusive'``,
+        ``'inner'``, ``'right'`` and ``'right_exclusive'``.
+    match_rows
+        Whether to match the indices of the element and table and if so how. If ``'left'``, element_indices take
+        priority and if ``'right'`` table instance ids take priority.
+
+    Returns
+    -------
+    A tuple containing the joined elements as a dictionary and the joined table as an AnnData object.
+
+    Raises
+    ------
+    AssertionError
+        If no table with the given table_name exists in the SpatialData object.
+    ValueError
+        If the provided join type is not supported.
+    """
+    assert sdata.tables.get(table_name), f"No table with `{table_name}` exists in the SpatialData object."
+    table = sdata.tables[table_name]
+    if isinstance(spatial_element_name, str):
+        spatial_element_name = [spatial_element_name]
+
+    elements_dict: dict[str, dict[str, Any]] = defaultdict(lambda: defaultdict(dict))
+    for name in spatial_element_name:
+        if name in sdata.tables:
+            warnings.warn(
+                f"Tables: `{', '.join(elements_dict['tables'].keys())}` given in spatial_element_names cannot be "
+                f"joined with a table using this function.",
+                UserWarning,
+                stacklevel=2,
+            )
+        elif name in sdata.images:
+            warnings.warn(
+                f"Images: `{', '.join(elements_dict['images'].keys())}` cannot be joined with a table",
+                UserWarning,
+                stacklevel=2,
+            )
+        else:
+            element_type, _, element = sdata._find_element(name)
+            elements_dict[element_type][name] = element
+
+    assert any(key in elements_dict for key in ["labels", "shapes", "points"]), (
+        "No valid element to join in spatial_element_name. Must provide at least one of either `labels`, `points` or "
+        "`shapes`."
+    )
+
+    if match_rows not in MatchTypes.__dict__["_member_names_"]:
+        raise TypeError(
+            f"`{match_rows}` is an invalid argument for `match_rows`. Can be either `no`, ``'left'`` or ``'right'``"
+        )
+    if how in JoinTypes.__dict__["_member_names_"]:
+        elements_dict, table = JoinTypes[how](elements_dict, table, match_rows)
+    else:
+        raise TypeError(f"`{how}` is not a valid type of join.")
+
+    elements_dict = {
+        name: element for outer_key, dict_val in elements_dict.items() for name, element in dict_val.items()
+    }
+    return elements_dict, table
+
+
 def match_table_to_element(sdata: SpatialData, element_name: str) -> AnnData:
     """
     Filter the table and reorders the rows to match the instances (rows/labels) of the specified SpatialElement.
@@ -138,17 +542,51 @@ def match_table_to_element(sdata: SpatialData, element_name: str) -> AnnData:
     sdata
         SpatialData object
     element_name
-        Name of the element to match the table to
+        The name of the spatial elements to be joined with the table.
 
     Returns
     -------
     Table with the rows matching the instances of the element
     """
+    # TODO: refactor this to make use of the new join_sdata_spatialelement_table function.
+    # if table_name is None:
+    #     warnings.warn(
+    #         "Assumption of table with name `table` being present is being deprecated in SpatialData v0.1. "
+    #         "Please provide the name of the table as argument to table_name.",
+    #         DeprecationWarning,
+    #         stacklevel=2,
+    #     )
+    #     table_name = "table"
+    # _, table = join_sdata_spatialelement_table(sdata, element_name, table_name, "left", match_rows="left")
+    # return table
     assert sdata.table is not None, "No table found in the SpatialData"
     element_type, _, element = sdata._find_element(element_name)
     assert element_type in ["labels", "shapes"], f"Element {element_name} ({element_type}) is not supported"
     elements_dict = {element_type: {element_name: element}}
     return _filter_table_by_elements(sdata.table, elements_dict, match_rows=True)
+
+
+def match_element_to_table(
+    sdata: SpatialData, element_name: str | list[str], table_name: str
+) -> tuple[dict[str, Any], AnnData]:
+    """
+    Filter the elements and make the indices match those in the table.
+
+    Parameters
+    ----------
+    sdata
+       SpatialData object
+    element_name
+       The name(s) of the spatial elements to be joined with the table. Not supported for Label elements.
+    table_name
+       The name of the table to join with the spatial elements.
+
+    Returns
+    -------
+    A tuple containing the joined elements as a dictionary and the joined table as an AnnData object.
+    """
+    element_dict, table = join_sdata_spatialelement_table(sdata, element_name, table_name, "right", match_rows="right")
+    return element_dict, table
 
 
 @dataclass